rename _help_files to help_files.  Move test code out of help_files  Remove help doc gen.
