+++ azure-cli.pyproj
@@ -287,8 +287,7 @@
     <Compile Include="azure\cli\tests\test_help.py" />
     <Compile Include="azure\cli\tests\test_output.py" />
     <Compile Include="azure\cli\_help.py" />
-    <Compile Include="azure\cli\_helpdocgen.py" />
-    <Compile Include="azure\cli\_help_files.py" />
+    <Compile Include="azure\cli\help_files.py" />
     <Compile Include="azure\cli\tests\test_profile.py" />
     <Compile Include="azure\cli\_logging.py" />
     <Compile Include="azure\cli\_output.py" />
